\name{tutorialsPascuzziPE}
\alias{tutorialsPascuzziPE}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
LearnR Tutorials by P.E. Pascuzzi
}
\description{
These learnR tutorials are a work in progress.  They were written to support both undergraduate and graduate courses for students in the life sciences.  The package can be installed or download from GitHub,  https://github.com/pepascuzzi/tutorialsPascuzziPE.  There are currently four tutorials.

    "T01_GettingStarted"

    "T02_DataTypes"

    "T03_UsingVectors"

    "T04_Tibbles"

    "T05_IntroToGgplot2"
}
\usage{
learnr::run_tutorial("tutorial name", package="tutorialsPascuzziPE")
}
\arguments{
  \item{x}{
}
}
\details{
The learnr and tidyverse packages must be installed and loaded.
}
\value{
not applicable
}
\references{
None at this time.
}
\author{
Pete E. Pascuzzi
}
\note{
None at this time.
}



\seealso{
learnR package
}
\examples{
library(learnr)
run_tutorial("T01_GettingStarted", "tutorialsPascuzziPE")
run_tutorial("T02_DataTypes", "tutorialsPascuzziPE")
run_tutorial("T03_UsingVectors", "tutorialsPascuzziPE")
run_tutorial("T04_Tibbles", "tutorialsPascuzziPE")
run_tutorial("T05_IntroToGgplot2", "tutorialsPascuzziPE")
run_tutorial("T06_DataWrangling01", "tutorialsPascuzziPE")
run_tutorial("T07_FunctionsAndLapply", "tutorialsPascuzziPE")
}

\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
